backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt>=5.0.0",
    "fastapi>=0.131.0",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
